chore: rename location to args
