+++ cmd/soroban-cli/src/config/identity/add.rs
@@ -1,4 +1,4 @@
-use crate::config::{location, secret};
+use crate::config::{args, secret};
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
@@ -6,7 +6,7 @@ pub enum Error {
     Secret(#[from] secret::Error),
 
     #[error(transparent)]
-    Config(#[from] location::Error),
+    Config(#[from] args::Error),
 
     #[error("Failed to write identity file")]
     IdCreationFailed,
@@ -25,7 +25,7 @@ pub struct Cmd {
     pub default: bool,
 
     #[clap(flatten)]
-    pub config: location::Args,
+    pub config: args::Args,
 }
 
 impl Cmd {
